Accion y vista editar por ID

+++ src/main/java/com/productosThymeleaft/repositorio/productosRepositorio.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface productosRepositorio extends JpaRepository<Productos,Integer> {
+public interface productosRepositorio extends JpaRepository<Productos,Long> {
 }
